+++ CHANGELOG.md
@@ -38,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed the issue where pure text can not be translated correctly. (#511, #514)
 - Fixed the issue where reading status not updating correctly in image gallery mode.
 - Fixed the issue where translation still happens when translation is disabled. (#541)
+- Fixed the issue where custom headers can input non-ascii characters. (#549)
 
 ## [1.3.16] - 2026-01-15